Исправление денег на BigDecimal и удаление лишнего импорта

+++ src/main/java/com/example/demo/model/db/entity/ElectricityPay.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -25,7 +26,7 @@ public class ElectricityPay {
     Plot plot;
 
     @Column(name = "amount")
-    Long amount;
+    BigDecimal amount;
 
     @Column(name = "date")
     LocalDateTime date;
